refactor: break out general concepts from database_connection and remove it from package exports.
