+++ packages/serverpod/lib/src/database/database.dart
@@ -3,6 +3,9 @@ import 'dart:async';
 import 'package:postgres/postgres.dart';
 import 'package:retry/retry.dart';
 import 'package:serverpod/src/database/columns.dart';
+import 'package:serverpod/src/database/includes.dart';
+import 'package:serverpod/src/database/order.dart';
+import 'package:serverpod/src/database/transaction.dart';
 
 import '../server/session.dart';
 import 'database_connection.dart';
